+++ src/commonTest/kotlin/electionguard/protoconvert/ElectionInitializedConvertTest.kt
@@ -63,7 +63,7 @@ fun generateElectionInitialized(context: GroupContext): ElectionInitialized {
 private fun generateGuardian(seq: Int, context: GroupContext): Guardian {
     return Guardian(
         "guardian $seq",
-        (seq + 1).toUInt(),
+        (seq + 1),
         List(3) { generateElementModP(context) },
         List(3) { generateSchnorrProof(context) },
     )
